requirements.txt
alabaster==0.7.12
appdirs==1.4.3
argon2-cffi==19.1.0
astroid==2.2.5
atomicwrites==1.3.0
attrs==19.1.0
Babel==2.6.0
backcall==0.1.0
black==19.3b0
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
Click==7.0
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5.3
decorator==4.4.0
defusedxml==0.6.0
Django==2.1.8
django-allauth==0.39.1
django-cors-headers==3.0.1
django-coverage-plugin==1.6.0
django-crispy-forms==1.7.2
django-debug-toolbar==1.11
django-environ==0.4.5
django-extensions==2.1.7
django-filter==2.1.0
django-model-utils==3.1.2
django-redis==4.10.0
django-rest-auth==0.9.5
djangorestframework==3.9.4
djangorestframework-jwt==1.11.0
docutils==0.14
docutils-stubs==0.0.21
entrypoints==0.3
factory-boy==2.12.0
Faker==1.0.7
flake8==3.7.7
idna==2.8
imagesize==1.1.0
ipdb==0.12
ipython==7.5.0
ipython-genutils==0.2.0
isort==4.3.20
itypes==1.1.0
jedi==0.13.3
Jinja2==2.10.1
lazy-object-proxy==1.4.1
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==7.0.0
mypy==0.701
mypy-extensions==0.4.1
oauthlib==3.0.1
packaging==19.0
parso==0.4.0
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.0.0
pluggy==0.11.0
prompt-toolkit==2.0.9
psycopg2-binary==2.8.2
ptyprocess==0.6.0
py==1.8.0
pycodestyle==2.5.0
pycparser==2.19
pyflakes==2.1.1
Pygments==2.4.1
PyJWT==1.7.1
pylint==2.3.1
pylint-django==2.0.9
pylint-plugin-utils==0.5
pyparsing==2.4.0
pytest==4.5.0
pytest-django==3.4.8
pytest-sugar==0.9.2
python-dateutil==2.8.0
python-slugify==3.0.2
python3-openid==3.1.0
pytz==2019.1
redis==3.2.1
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
snowballstemmer==1.2.1
Sphinx==2.0.1
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
sqlparse==0.3.0
termcolor==1.1.0
text-unidecode==1.2
toml==0.10.0
traitlets==4.3.2
typed-ast==1.3.5
uritemplate==3.0.0
urllib3==1.25.3
wcwidth==0.1.7
Werkzeug==0.14.1
wrapt==1.11.1
